Turn State into a trait
Convert HikariDataSource into first implementation
