+++ core/src/main/scala/ai/lum/odinson/state/StateFactory.scala
@@ -1,0 +1,15 @@
+package ai.lum.odinson.state
+
+import com.typesafe.config.Config
+
+import ai.lum.common.ConfigUtils._
+
+object StateFactory {
+
+  def newState(config: Config): State = {
+    val jdbcUrl = config[String]("state.jdbc.url")
+    val state = new HikariState(jdbcUrl)
+
+    state
+  }
+}
